+++ src/Firebase/Messaging/Http/Request/SendMessageToTokens.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Messaging\Http\Request;
 
+use Kreait\Firebase\Contract\Messaging;
 use Kreait\Firebase\Exception\Messaging\InvalidArgument;
 use Kreait\Firebase\Http\HasSubRequests;
 use Kreait\Firebase\Http\WrappedPsr7Request;
@@ -20,12 +21,15 @@ final class SendMessageToTokens implements HasSubRequests, RequestInterface
 {
     use WrappedPsr7Request;
 
-    public const MAX_AMOUNT_OF_TOKENS = 500;
+    /**
+     * @deprecated 6.6.0 Use {@see Messaging::BATCH_MESSAGE_LIMIT} instead
+     */
+    public const MAX_AMOUNT_OF_TOKENS = Messaging::BATCH_MESSAGE_LIMIT;
 
     public function __construct(string $projectId, Message $message, RegistrationTokens $registrationTokens, bool $validateOnly = false)
     {
-        if ($registrationTokens->count() > self::MAX_AMOUNT_OF_TOKENS) {
-            throw new InvalidArgument('A multicast message can be sent to a maximum amount of '.self::MAX_AMOUNT_OF_TOKENS.' tokens.');
+        if ($registrationTokens->count() > Messaging::BATCH_MESSAGE_LIMIT) {
+            throw new InvalidArgument('A multicast message can be sent to a maximum amount of '.Messaging::BATCH_MESSAGE_LIMIT.' tokens.');
         }
 
         $messageData = $message->jsonSerialize();